fix: resolve mypy type checking error in embedding backend
Fix incompatible type assignment in speaker identification method:

• Convert numpy.mean() result to Python float to match variable type
• Resolve mypy error: "expression has type 'floating[Any]', variable has type 'float'"
• Line 213: avg_similarity = float(np.mean(similarities))
• Maintain full functionality while ensuring type safety
• All mypy type checks now pass cleanly

Speaker recognition system now passes all static analysis checks.

+++ src/speaker/embedding_backend.py
@@ -210,7 +210,7 @@ class EmbeddingSpeakerBackend:
 
                 # Use average similarity across all samples
                 if similarities:
-                    avg_similarity = np.mean(similarities)
+                    avg_similarity = float(np.mean(similarities))
                     if avg_similarity > best_similarity and avg_similarity >= user_threshold:
                         best_similarity = avg_similarity
                         best_user = user_id
